Fix #14723: Set force proceed to TFP_SIGNAL if next tile has signals.
Old behavior was to set `t->force_proceed` to TFP_STUCK whenever the train is marked as stuck.
But when train stops before signal it is also marked as stuck.
Setting `t->force_proceed` to TFP_STUCK in such situation does not allow train to proceed,
because it needs `t->force_proceed == TFP_SIGNAL` in order to skip red/oneway signals.
Checking if the next tile has signals is nearly ideal way to
check whether the train is stuck because of signal or not.